Typos

+++ src/sfm/triangulation.h
@@ -7,6 +7,6 @@
 
 void triangulate(VOFrame *vo0, VOFrame *vo1);
 
-double getScale(const VOFrame &vo0, const VOFrame &vo1, int min_points, int max_points);
+float getScale(const VOFrame &vo0, const VOFrame &vo1, int min_points, int max_points);
 
 #endif //VO_TRIANGULATION_H
